geolocation/geo-clearWatch-005.html

<!DOCTYPE html>
<html>
  <head>
    <title>Geolocation Test: call geolocation.clearWatch with empty arguments, exception is thrown</title>
    <link rel="author" title="Intel" href="http://www.intel.com/">
    <link rel="help" href="http://www.w3.org/TR/2012/PR-geolocation-API-20120510/#geolocation">
    <link rel="help" href="http://www.w3.org/TR/2012/PR-geolocation-API-20120510/#watch-position">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
  <div id="log"></div>
    <script>
      var t = async_test("Test checks that call geolocation.clearWatch with empty arguments, exception is thrown");
      setup({timeout: 20000});
      try {
        navigator.geolocation.clearWatch();
      } catch (err) {
        t.step(function() {
          assert_equals(err.message, "Not enough arguments", "call geolocation.clearWatch with no arguments, exception is thrown");
        });
        t.done();
      }
    </script>
  </body>
</html>
